backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-SQLAlchemy==3.1.1
Flask-Talisman==1.1.0
Werkzeug==3.0.1
APScheduler==3.10.4
reportlab==4.0.8
openai==1.6.1
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
email-validator==2.1.0.post1
pytest==7.4.3